Beginning to add infra for Characters...
Characters will represent how a creature came to be over history,
encoding things like overall race, levels over time, and whatever that
character is wearing. They will be the unit of creature import/export.

BaseCreature will still be the living representation of a creature as it
is "right now", especially for combat simulation.